% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{otter_download_pdf_report}
\alias{otter_download_pdf_report}
\title{PDF report}
\usage{
otter_download_pdf_report(otter_obj, inference_id, output_path)
}
\arguments{
\item{otter_obj}{Your API object.}

\item{inference_id}{Inference ID of the sample.}

\item{output_path}{Path to save the pdf.}
}
\value{
PDF file.
}
\description{
PDF report of the sample.
}
