% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{otter_explore_plots}
\alias{otter_explore_plots}
\title{Explore plots}
\usage{
otter_explore_plots(otter_obj, class_name = "")
}
\arguments{
\item{otter_obj}{Your API object.}

\item{class_name}{Optional parameter to get the explore plots of an specific class.}
}
\value{
Plotly plot.
}
\description{
Plots available in the explore page.
}
