% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{otter_plot_sample}
\alias{otter_plot_sample}
\title{Main plot}
\usage{
otter_plot_sample(otter_obj, df_result, width = 800, height = 800)
}
\arguments{
\item{otter_obj}{Your API object.}

\item{df_result}{Dataframe of prediction.}

\item{width}{Plot width.}

\item{height}{Plot height.}
}
\value{
Plotly plot.
}
\description{
Sunburst plot and other plots and info.
}
