% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{otter_to_plot}
\alias{otter_to_plot}
\title{Local plots}
\usage{
otter_to_plot(plot_data)
}
\arguments{
\item{plot_data}{Result from the plotting functions from otter.}
}
\value{
Plots.
}
\description{
Use plotly to create local plots.
}
