% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{otter_wait_for_sample}
\alias{otter_wait_for_sample}
\title{Waits for the API}
\usage{
otter_wait_for_sample(otter_obj, task_id, timeout)
}
\arguments{
\item{otter_obj}{Your API object.}

\item{task_id}{The id of the API run.}

\item{timeout}{Time to timeout.}
}
\value{
Prediction dataframe
}
\description{
Waits for the results to be available in the API
}
